md_a2a_v2/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
httpx>=0.24.0  # For TestClient